+++ PyMca5/PyMcaGraph/ctools/_ctools/cython/MinMax.pxd
@@ -1,7 +1,7 @@
 cdef extern from "MinMax.h":
     void getMinMax(void * data,
                    unsigned int type,
-                   unsigned int length,
+                   unsigned long length,
                    double * minOut,
                    double * minPosOut,
                    double * maxOut) nogil
